+++ packages/client-core/src/systems/WarningUISystem.tsx
@@ -33,6 +33,7 @@ import { Engine } from '@ir-engine/ecs/src/Engine'
 import { removeEntity } from '@ir-engine/ecs/src/EntityFunctions'
 import { defineSystem } from '@ir-engine/ecs/src/SystemFunctions'
 import { PresentationSystemGroup } from '@ir-engine/ecs/src/SystemGroups'
+import { createXRUI } from '@ir-engine/engine/src/xrui/createXRUI'
 import { defineState, getMutableState, getState, useMutableState } from '@ir-engine/hyperflux'
 import { EngineState } from '@ir-engine/spatial/src/EngineState'
 import { CameraComponent } from '@ir-engine/spatial/src/camera/components/CameraComponent'
@@ -40,9 +41,8 @@ import { NameComponent } from '@ir-engine/spatial/src/common/NameComponent'
 import { createTransitionState } from '@ir-engine/spatial/src/common/functions/createTransitionState'
 import { VisibleComponent, setVisibleComponent } from '@ir-engine/spatial/src/renderer/components/VisibleComponent'
 import { ComputedTransformComponent } from '@ir-engine/spatial/src/transform/components/ComputedTransformComponent'
+import { ObjectFitFunctions } from '@ir-engine/spatial/src/transform/functions/ObjectFitFunctions'
 import { XRUIComponent } from '@ir-engine/spatial/src/xrui/components/XRUIComponent'
-import { ObjectFitFunctions } from '@ir-engine/spatial/src/xrui/functions/ObjectFitFunctions'
-import { createXRUI } from '@ir-engine/spatial/src/xrui/functions/createXRUI'
 import type { WebLayer3D } from '@ir-engine/xrui'
 
 export const WarningUIState = defineState({
