+++ packages/client-core/src/systems/createAnchorWidget.tsx
@@ -28,15 +28,15 @@ import { MathUtils } from 'three'
 
 import { getComponent, removeComponent } from '@ir-engine/ecs/src/ComponentFunctions'
 import { ECSState } from '@ir-engine/ecs/src/ECSState'
+import { createXRUI } from '@ir-engine/engine/src/xrui/createXRUI'
 import { dispatchAction, getMutableState, getState, startReactor, useHookstate } from '@ir-engine/hyperflux'
 import { InputSourceComponent } from '@ir-engine/spatial/src/input/components/InputSourceComponent'
 import { XRStandardGamepadAxes, XRStandardGamepadButton } from '@ir-engine/spatial/src/input/state/ButtonState'
 import { InputState } from '@ir-engine/spatial/src/input/state/InputState'
 import { VisibleComponent } from '@ir-engine/spatial/src/renderer/components/VisibleComponent'
 import { XRState } from '@ir-engine/spatial/src/xr/XRState'
-import { createXRUI } from '@ir-engine/spatial/src/xrui/functions/createXRUI'
-import { WidgetAppActions } from '@ir-engine/spatial/src/xrui/WidgetAppService'
-import { Widget, Widgets } from '@ir-engine/spatial/src/xrui/Widgets'
+import { WidgetAppActions } from './WidgetAppService'
+import { Widget, Widgets } from './Widgets'
 
 import { defineQuery } from '@ir-engine/ecs'
 import { AnchorWidgetUI } from './ui/AnchorWidgetUI'
