+++ packages/client-core/src/systems/createMediaWidget.tsx
@@ -23,8 +23,8 @@ All portions of the code written by the Infinite Reality Engine team are Copyrig
 Infinite Reality Engine. All Rights Reserved.
 */
 
-import { createXRUI } from '@ir-engine/spatial/src/xrui/functions/createXRUI'
-import { Widget, Widgets } from '@ir-engine/spatial/src/xrui/Widgets'
+import { createXRUI } from '@ir-engine/engine/src/xrui/createXRUI'
+import { Widget, Widgets } from './Widgets'
 
 import { UserMediaWindowsWidget } from '../components/UserMediaWindows'
