+++ packages/client-core/src/systems/createRecordingsWidget.tsx
@@ -24,9 +24,9 @@ Infinite Reality Engine. All Rights Reserved.
 */
 
 import { removeComponent } from '@ir-engine/ecs/src/ComponentFunctions'
+import { createXRUI } from '@ir-engine/engine/src/xrui/createXRUI'
 import { VisibleComponent } from '@ir-engine/spatial/src/renderer/components/VisibleComponent'
-import { createXRUI } from '@ir-engine/spatial/src/xrui/functions/createXRUI'
-import { Widget, Widgets } from '@ir-engine/spatial/src/xrui/Widgets'
+import { Widget, Widgets } from './Widgets'
 
 import { RecordingsWidgetUI } from './ui/RecordingsWidgetUI'
