+++ packages/client-core/src/systems/ui/AvatarDetailView/index.tsx
@@ -30,11 +30,11 @@ import { CircleGeometry, Mesh, MeshBasicMaterial } from 'three'
 import { useGet } from '@ir-engine/common'
 import { userPath } from '@ir-engine/common/src/schema.type.module'
 import { setComponent } from '@ir-engine/ecs/src/ComponentFunctions'
+import { createXRUI } from '@ir-engine/engine/src/xrui/createXRUI'
+import { useXRUIState } from '@ir-engine/engine/src/xrui/useXRUIState'
 import { getMutableState, hookstate, useHookstate } from '@ir-engine/hyperflux'
 import { NetworkState } from '@ir-engine/network'
 import { NameComponent } from '@ir-engine/spatial/src/common/NameComponent'
-import { createXRUI } from '@ir-engine/spatial/src/xrui/functions/createXRUI'
-import { useXRUIState } from '@ir-engine/spatial/src/xrui/functions/useXRUIState'
 
 import { AvatarUIState } from '../../state/AvatarUIState'
 import styleString from './index.scss?inline'
