+++ packages/client-core/src/systems/ui/LoadingDetailView/index.tsx
@@ -26,9 +26,9 @@ Infinite Reality Engine. All Rights Reserved.
 import React from 'react'
 import { useTranslation } from 'react-i18next'
 
+import { createXRUI } from '@ir-engine/engine/src/xrui/createXRUI'
+import { useXRUIState } from '@ir-engine/engine/src/xrui/useXRUIState'
 import { hookstate, State } from '@ir-engine/hyperflux'
-import { createXRUI } from '@ir-engine/spatial/src/xrui/functions/createXRUI'
-import { useXRUIState } from '@ir-engine/spatial/src/xrui/functions/useXRUIState'
 
 import ProgressBar from './SimpleProgressBar'
 import LoadingDetailViewStyle from './style'
