+++ packages/client-core/src/systems/ui/RecordingsWidgetUI.tsx
@@ -32,8 +32,8 @@ import { RecordingType, recordingPath } from '@ir-engine/common/src/schema.type.
 import { Engine } from '@ir-engine/ecs/src/Engine'
 import { PeerID, defineState, getMutableState, getState, useHookstate, useMutableState } from '@ir-engine/hyperflux'
 import { NetworkState } from '@ir-engine/network'
-import { WidgetAppService } from '@ir-engine/spatial/src/xrui/WidgetAppService'
 import { startPlayback } from '@ir-engine/ui/src/pages/Capture'
+import { WidgetAppService } from '../WidgetAppService'
 
 import { useMediaNetwork } from '../../common/services/MediaInstanceConnectionService'
 import { PeerMediaChannelState } from '../../media/PeerMediaChannelState'
