+++ packages/client-core/src/systems/ui/WidgetMenuView/index.tsx
@@ -28,12 +28,12 @@ import React from 'react'
 import { AvatarComponent } from '@ir-engine/engine/src/avatar/components/AvatarComponent'
 // import { VrIcon } from '../../../common/components/Icons/VrIcon'
 import { respawnAvatar } from '@ir-engine/engine/src/avatar/functions/respawnAvatar'
+import { createXRUI } from '@ir-engine/engine/src/xrui/createXRUI'
 import { createState, dispatchAction, getMutableState, useHookstate, useMutableState } from '@ir-engine/hyperflux'
 import { InputState } from '@ir-engine/spatial/src/input/state/InputState'
 import { XRState } from '@ir-engine/spatial/src/xr/XRState'
-import { createXRUI } from '@ir-engine/spatial/src/xrui/functions/createXRUI'
-import { RegisteredWidgets, WidgetAppActions, WidgetAppState } from '@ir-engine/spatial/src/xrui/WidgetAppService'
 import Icon from '@ir-engine/ui/src/primitives/mui/Icon'
+import { RegisteredWidgets, WidgetAppActions, WidgetAppState } from '../../WidgetAppService'
 
 import { useMediaInstance } from '../../../common/services/MediaInstanceConnectionService'
 import { MediaStreamState } from '../../../media/MediaStreamState'
