+++ packages/client-core/src/world/ClientModule.ts
@@ -23,7 +23,7 @@ All portions of the code written by the Infinite Reality Engine team are Copyrig
 Infinite Reality Engine. All Rights Reserved.
 */
 
-import { WidgetAppServiceReceptorSystem } from '@ir-engine/spatial/src/xrui/WidgetAppService'
+import { WidgetAppServiceReceptorSystem } from '../systems/WidgetAppService'
 
 import { AvatarSpawnSystem } from '../networking/AvatarSpawnSystem'
 // import { AvatarUISystem } from '../systems/AvatarUISystem'
