+++ packages/engine/src/xrui/createXRUI.tsx
@@ -35,16 +35,16 @@ import { WebContainer3D } from '@ir-engine/xrui/core/three/WebContainer3D'
 import { WebLayerManager } from '@ir-engine/xrui/core/three/WebLayerManager'
 
 import { AssetLoaderState } from '@ir-engine/engine/src/assets/state/AssetLoaderState'
-import { EngineState } from '../../EngineState'
-import { InputComponent } from '../../input/components/InputComponent'
-import { RendererComponent } from '../../renderer/WebGLRendererSystem'
-import { addObjectToGroup } from '../../renderer/components/GroupComponent'
-import { setObjectLayers } from '../../renderer/components/ObjectLayerComponent'
-import { VisibleComponent } from '../../renderer/components/VisibleComponent'
-import { ObjectLayers } from '../../renderer/constants/ObjectLayers'
-import { DistanceFromCameraComponent } from '../../transform/components/DistanceComponents'
-import { XRUIStateContext } from '../XRUIStateContext'
-import { XRUIComponent } from '../components/XRUIComponent'
+import { EngineState } from '@ir-engine/spatial/src/EngineState'
+import { InputComponent } from '@ir-engine/spatial/src/input/components/InputComponent'
+import { RendererComponent } from '@ir-engine/spatial/src/renderer/WebGLRendererSystem'
+import { addObjectToGroup } from '@ir-engine/spatial/src/renderer/components/GroupComponent'
+import { setObjectLayers } from '@ir-engine/spatial/src/renderer/components/ObjectLayerComponent'
+import { VisibleComponent } from '@ir-engine/spatial/src/renderer/components/VisibleComponent'
+import { ObjectLayers } from '@ir-engine/spatial/src/renderer/constants/ObjectLayers'
+import { DistanceFromCameraComponent } from '@ir-engine/spatial/src/transform/components/DistanceComponents'
+import { XRUIComponent } from '@ir-engine/spatial/src/xrui/components/XRUIComponent'
+import { XRUIStateContext } from './XRUIStateContext'
 
 export function createXRUI<S extends State<any> | null>(
   UIFunc: React.FC,
